00:17 Driect3D 12 calculation 2 KKH

+++ KKH/3D_ViewerProject/3D_ViewerProject/Const.h
@@ -10,6 +10,6 @@ const WORD WINSIZE_Y = 1050;
 const float SCREEN_DEPTH = 1000.0f;
 const float SCREEN_NEAR = 0.1f;
 
-const double PI = 3.14159265358979323846;
+const float PI = 3.14159265358979323846;
 
 const int NumFrameResources = 3;
